--- brainstorminng_behavython.txt
@@ -1,50 +1,0 @@
-
-
-- Desabilitar os botões de análise enquanto a experimentador não selecionar o config.yaml e a pasta com os vídeos
-- Se a experimentador não selecionar o arquivo correto (config.yaml), avisar.
-- Só deixar a análise correr se a pasta da rede estiver OK
-- *Inverter as abas analysis e deeplabcut* para refletir a ordem de análise
-- Mostrar um popup enquanto a análise do deeplabcut está correndo para a experimentador não achar que o programa travou
-	- Mostrar que depende das configurações do computador e que é imprecindível a presença de uma placa de vídeo
-	- E avisar sobre a quantidade de vídeos
-- Alterar a função que extrai os frames para que a extração seja feita de maneira dinâmica (relativo ao tamanho. Ex.: 2/3)
-
-- Criar uma sequência de destrave na interface, seguindo a sequência de análise.
-- Adicionar um botão (standalone) que passa automaticamente por toda análise sem a necessidade de interação do experimentador.
-
-- Alterar os labels para que, ao passar o mouse em cima, alterar o ícone para "mouse com pergunta" e, ao clicar, aparece
-uma descrição mais detalhada
-- Colocar um botão de informação para que o experimentador clique e seja levado a um popup.
-
-- Colocar um popup explicando que o experimentador precisa extrair os ROIs (Region of interest) utilizando o ImageJ.
-- Trocar D e E por R e L  ----DONE----
-
-- (BUG) Quando a imagem está no diretório o programa apaga seus dados. ----DONE----
-
-- Trocar GET FRAMES para GET REPRESENTATIVE FRAME (ou GET FRAME) ----DONE----
-
-- Checar quais arquivos estão presentes no diretório e comparar com qual tipo de análise está tentando ser feita. 
-Se estiver errado, não deixar o experimentador realizar a análise/avisar
-
-- Padronizar os plots do bonsai e do deeplabcut
-
-- No início da análise, avisar para que o experimentador confira BEEEEM os dados antes de prosseguir porque esses são muito
-importantes para o resultado:
-	- Tamanho da arena
-	- Frames por segundo
-	- Duração da tarefa
-	- Animal experimental
-	
-- Conferir se, o experimentador optar por cortar o vídeo, o tempo para ser aparado (trim) é ser menor do que a duração da 
-tarefa
-
-- Mudar o botão de crop video para settings e trocar o padrão ou para radiobutton ou para o atual
-
-- Adicionar uma mensagem que mostra se o vídeo tem tempo suficiente para que a tarefa seja concluída
-	Exemplo: Se eu coloco na interface que o tempo de duração da tarefa tem 300 segundos, o vídeo precisa ter 300 segundos mais o tempo extra para ser removido (caso seja marcada essa opção). Então 300s + 15s, por exemplo.
-	Checar isso no código.
-	Caso exista algum erro em algum vídeo, não parar a análise, só pular o vídeo e informar depois.
-	
-
-- Se o experimentador desejar somente realizar a análise de arquivos já feitos (provindos do DLC), desabilitar o import 
-do deeplabcut para acelerar a inicialização.
