--- network/CD1-Matheus-2023-02-12/dlc-models/iteration-1/CD1Feb12-trainset95shuffle1/test/pose_cfg.yaml
@@ -1,18 +1,0 @@
-all_joints:
-- - 0
-- - 1
-- - 2
-- - 3
-all_joints_names:
-- Focinho
-- OrelhaE
-- OrelhaD
-- Rabo
-dataset: training-datasets\iteration-1\UnaugmentedDataSet_CD1Feb12\CD1_Matheus95shuffle1.mat
-global_scale: 0.8
-init_weights: C:\Users\uzuna\mambaforge\envs\DLC\lib\site-packages\deeplabcut\pose_estimation_tensorflow\models\pretrained\mobilenet_v2_1.0_224.ckpt
-location_refinement: true
-locref_stdev: 7.2801
-net_type: mobilenet_v2_1.0
-num_joints: 4
-scoremap_dir: test
