--- network/CD1-Matheus-2023-02-12/dlc-models/iteration-1/CD1Feb12-trainset95shuffle1/train/pose_cfg.yaml
@@ -1,70 +1,0 @@
-all_joints:
-- - 0
-- - 1
-- - 2
-- - 3
-all_joints_names:
-- Focinho
-- OrelhaE
-- OrelhaD
-- Rabo
-alpha_r: 0.02
-apply_prob: 0.5
-batch_size: 1
-contrast:
-  clahe: true
-  claheratio: 0.1
-  histeq: true
-  histeqratio: 0.1
-convolution:
-  edge: false
-  emboss:
-    alpha:
-    - 0.0
-    - 1.0
-    strength:
-    - 0.5
-    - 1.5
-  embossratio: 0.1
-  sharpen: false
-  sharpenratio: 0.3
-cropratio: 0.4
-dataset: training-datasets\iteration-1\UnaugmentedDataSet_CD1Feb12\CD1_Matheus95shuffle1.mat
-dataset_type: imgaug
-decay_steps: 30000
-display_iters: 100
-global_scale: 0.8
-init_weights: C:\Users\uzuna\mambaforge\envs\DLC\lib\site-packages\deeplabcut\pose_estimation_tensorflow\models\pretrained\mobilenet_v2_1.0_224.ckpt
-intermediate_supervision: false
-intermediate_supervision_layer: 12
-location_refinement: true
-locref_huber_loss: true
-locref_loss_weight: 0.05
-locref_stdev: 7.2801
-lr_init: 0.0005
-max_input_size: 1500
-metadataset: training-datasets\iteration-1\UnaugmentedDataSet_CD1Feb12\Documentation_data-CD1_95shuffle1.pickle
-min_input_size: 64
-mirror: false
-multi_stage: false
-multi_step:
-- - 0.005
-  - 10000
-- - 0.02
-  - 430000
-- - 0.002
-  - 730000
-- - 0.001
-  - 1030000
-net_type: mobilenet_v2_1.0
-num_joints: 4
-pairwise_huber_loss: false
-pairwise_predict: false
-partaffinityfield_predict: false
-pos_dist_thresh: 17
-project_path: D:\DLC_test\CD1-Matheus-2023-02-12
-rotation: 25
-rotratio: 0.4
-save_iters: 1000
-scale_jitter_lo: 0.5
-scale_jitter_up: 1.25
