--- network/CD1-Matheus-2023-02-12/videos/above_roi.csv
@@ -1,2 +1,0 @@
- ,Area,Min,Max,X,Y,BX,BY,Width,Height,%Area
-1,31757,14,33,296.500,261.500,196,161,201,201,0
